+++ data_reading/wsfr_read/streamflow/__init__.py
@@ -1,5 +1,7 @@
+from wsfr_read.streamflow.naturalized_flow import read_test_monthly_naturalized_flow
 from wsfr_read.streamflow.usgs_streamflow import read_usgs_streamflow_data
 
 __all__ = [
+    "read_test_monthly_naturalized_flow",
     "read_usgs_streamflow_data",
 ]